fintel/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fintel-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "redis[hiredis]>=5.0.0",
    "celery[redis]>=5.3.0",
    "flower>=2.0.0",
    "httpx>=0.27.0",
    "playwright>=1.43.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "tenacity>=8.3.0",
    "python-dotenv>=1.0.0",
    "pgvector>=0.3.0",
    "openai>=1.30.0",
    "duckdb>=1.0.0",
    "pypdf>=4.2.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]
